web/dev-server/pyproject.toml
[project]
name = "gondul-dev-server"
version = "0.1.0"
description = "Gondul frontend Development server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
]
